refactor: change mane function
