+++ css-selectors/src/components/game/task-complite.ts
@@ -1,5 +1,5 @@
 import { DataStorage } from '../classes/Storage';
-import { createHtmlEditor } from '../view/create-next-html-editor';
+import { createHtmlEditor } from '../view/create-next_page';
 import { highlightCode } from '../view/highlight-code';
 import { removeElementPrevLevel } from '../view/remove-element';
 import { storageAnswers } from './answers';
